requirements.txt
beautifulsoup4==4.12.2
face-recognition==1.3.0
faiss-cpu==1.7.4
numpy==1.26.4
opencv-python==4.8.1.78
requests==2.31.0
tqdm==4.66.1
rich==13.7.0
